+++ py/desimodel/io.py
@@ -12,22 +12,6 @@ import yaml
 import numpy as np
 import warnings
 
-#
-#- PSF and throughput, which require specter
-#
-try:
-    from specter.throughput import load_throughput as specter_load_throughput
-except ImportError as e:
-    warnings.warn(str(e))
-    warnings.warn("Unable to import specter.throughput.load_throughput(); desimodel.io.load_throughput() won't work.")
-try:
-    from specter.psf import load_psf as specter_load_psf
-except ImportError as e:
-    warnings.warn(str(e))
-    warnings.warn("Unable to import specter.psf.load_psf(); desimodel.io.load_psf() won't work.")
-#
-#
-#
 _thru = dict()
 def load_throughput(channel):
     """Returns specter Throughput object for the given channel 'b', 'r', or 'z'.
@@ -37,6 +21,7 @@ def load_throughput(channel):
     channel : {'b', 'r', 'z'}
         Spectrograph channel.
     """
+    from specter.throughput import load_throughput as specter_load_throughput    
     channel = channel.lower()
     global _thru
     if channel not in _thru:
@@ -55,6 +40,7 @@ def load_psf(channel):
     channel : {'b', 'r', 'z'}
         Spectrograph channel.
     """
+    from specter.psf import load_psf as specter_load_psf
     channel = channel.lower()
     global _psf
     if channel not in _psf: